+++ CHANGELOG.md
@@ -6,11 +6,22 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking Changes
 
+- Moved rows plugin from `@edtr-io/ui` to `@edtr-io/plugin-rows`
 - Replaced `showOverlay` and `hideOverlay` with an `OverlayContext`
 
 ### Added
 
-- Added `InlineOverlay` Component for plugin controls
+- Added `InlineOverlay` component for plugin controls
+- Added focus previous / focus next actions
+- Added optional `getFocusableChildren` to stateful plugins, which is used to resolve the previous / next focusable child
+
+### Changed
+
+- **plugin-text**. Handle focus when using arrow keys
+
+### Fixed
+
+- **plugin-text**. Focus / blur slate depending on `props.focused`
 
 ## [0.3.2](https://github.com/edtr-io/edtr-io/compare/0.3.1..0.3.2) - March 14, 2019
 
@@ -18,7 +29,7 @@ All notable changes to this project will be documented in this file.
 
 - Added Image plugin
 - Added `editable` prop to `Editor` for switching between edit and render mode
-- Added optional `onPaste` callback to Plugins, which is called by the text plugin on paste event
+- Added optional `onPaste` callback to plugins, which is called by the text plugin on paste event
 
 ## [0.3.1](https://github.com/edtr-io/edtr-io/compare/0.3.0..0.3.1) - March 13, 2019